added style file to mood
