+++ Aura_frontend/app/(tabs)/Mood/Mood.tsx
@@ -4,11 +4,11 @@ import {
   Text,
   TouchableOpacity,
   Image,
-  StyleSheet,
   Alert,
 } from "react-native";
+import styles from "./Mood.styles";
 import Modal from "react-native-modal";
-import { logMood } from "../services/moodService";
+import { logMood } from "../../services/moodService";
 
 //define the props for moodlog
 type MoodLogProps = {
@@ -18,22 +18,20 @@ type MoodLogProps = {
 
 //list of mood images to select
 const moods = [
-  { label: "Happy", image: require("../../assets/images/happy.png") },
-  { label: "Neutral", image: require("../../assets/images/neutral.png") },
-  { label: "Sad", image: require("../../assets/images/sad.png") },
-  { label: "Depressed", image: require("../../assets/images/depressed.png") },
-  { label: "Angry", image: require("../../assets/images/angry.png") },
+  { label: "Happy", image: require("../../../assets/images/happy.png") },
+  { label: "Neutral", image: require("../../../assets/images/neutral.png") },
+  { label: "Sad", image: require("../../../assets/images/sad.png") },
+  { label: "Depressed", image: require("../../../assets/images/depressed.png") },
+  { label: "Angry", image: require("../../../assets/images/angry.png") },
 ];
 
 const MoodLog: React.FC<MoodLogProps> = ({ isVisible, onClose }) => {
-
   const [selectedMood, setSelectedMood] = useState<string | null>(null); // tracking the state of the selected mood
 
   //handle mood submission
   const handleSubmitMood = async () => {
-    
     // if no mood selected, prevent the submission
-    if (!selectedMood) { 
+    if (!selectedMood) {
       Alert.alert(
         "No Mood Selected",
         "Please select a mood before proceeding."
@@ -43,7 +41,7 @@ const MoodLog: React.FC<MoodLogProps> = ({ isVisible, onClose }) => {
 
     try {
       await logMood(selectedMood); //call moodService to add mood
-      
+
       // show sucess message & close the modal
       Alert.alert(
         "Mood Noted",
@@ -68,8 +66,6 @@ const MoodLog: React.FC<MoodLogProps> = ({ isVisible, onClose }) => {
       animationIn="slideInUp"
       animationOut="slideOutDown"
     >
-
-
       {/* Mood selector pop up */}
       <View style={styles.container}>
         <View style={styles.swipeBar} />
@@ -81,7 +77,6 @@ const MoodLog: React.FC<MoodLogProps> = ({ isVisible, onClose }) => {
 
         {/* horizontal row of moods */}
         <View style={styles.moodRow}>
-
           {moods.map((mood) => (
             <TouchableOpacity
               key={mood.label}
@@ -91,13 +86,11 @@ const MoodLog: React.FC<MoodLogProps> = ({ isVisible, onClose }) => {
               ]}
               onPress={() => setSelectedMood(mood.label)} // Set selected mood
             >
-              <Image source={mood.image} style={styles.moodImage} /> 
+              <Image source={mood.image} style={styles.moodImage} />
               <Text style={styles.moodLabel}>{mood.label}</Text>
             </TouchableOpacity>
           ))}
-
         </View>
-
 
         {/* button */}
         <TouchableOpacity style={styles.button} onPress={handleSubmitMood}>
@@ -108,69 +101,5 @@ const MoodLog: React.FC<MoodLogProps> = ({ isVisible, onClose }) => {
   );
 };
 
-const styles = StyleSheet.create({
-  modal: {
-    justifyContent: "flex-end",
-    margin: 0,
-  },
-  container: {
-    backgroundColor: "#DFF3E4",
-    padding: 20,
-    borderTopLeftRadius: 25,
-    borderTopRightRadius: 25,
-  },
-  swipeBar: {
-    width: 40,
-    height: 5,
-    backgroundColor: "#ccc",
-    alignSelf: "center",
-    borderRadius: 3,
-    marginBottom: 10,
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: "bold",
-    marginBottom: 20,
-  },
-  moodRow: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    marginBottom: 20,
-  },
-  moodItem: {
-    alignItems: "center",
-    width: 60,
-    padding: 5,
-    borderRadius: 10,
-  },
-  selectedMood: {
-    backgroundColor: "#A7E9AF",
-  },
-  moodImage: {
-    width: 40,
-    height: 40,
-    marginBottom: 5,
-  },
-  moodLabel: {
-    fontSize: 9,
-  },
-  moodQuestion: {
-    fontSize: 15,
-    fontWeight: '500',
-    marginBottom: 15,
-    textAlign: "center",
-  },
-  button: {
-    backgroundColor: "#4CAF50",
-    paddingVertical: 12,
-    borderRadius: 25,
-    alignItems: "center",
-  },
-  buttonText: {
-    color: "#fff",
-    fontWeight: "600",
-    fontSize: 16,
-  },
-});
 
 export default MoodLog;
